NetApp ONTAP: Fix type error exception in get_volume_state
When a Flexgroup is being replicated in ONTAP, a snapmirror is created.
However, while ensuring the snapmirror creation, a type error occurs.

The function get_volume_state has a parameter name set as 'name', and it
is called under 'flexvol_name' causing the type error exception.

This patch solves the bug by fixing 'flexvol_name' to 'name'

Closes-Bug: #1958245
Change-Id: Ibf9823a47980901042a9c95ca3cd850cd14f0033