[Lightos] standardize connector usage
Lightos connector created a new key in the connection_properties dict
called "hostnqn" this key is a duplication of the "nqn" key created
in "nvmeof" connector.
This commit remove the use of "hostnqn" for "nqn".

There was a use of the key "nqn" in a different context and it
was changed: "nqn"-> "subsysnqn"

The driver had used a connector function outside the standard
connector interface.
It now gets the info supplied by that function from the standard
get_connector_properties().

Remove 'nqn' key from the connection_properties dict

This patch is a followup to change I2e86fa840490.

Signed-off-by: Yuval Brave  <yuval@lightbitslabs.com>
Depends-On: Ia4d8c6fda875b4f4d32d511ca9282ca0fb5d6f12
Change-Id: I68f6cb9a7da56966827bb8683da44e6852549cc4