RBD: Fix total_capacity
Ceph has changed the meaning of the ``bytes_used`` column in the pools
reported by the ``df`` command, which means that in some deployments the
rbd driver is not reporting the expected information ot the schedulers.

The information we should be used for the calculations is returned in
the ``stored`` field in those systems.

This patch uses ``stored`` when present and fallbacks to ``bytes_used``
if not.

Closes-Bug: #1960206
Change-Id: I0ca25789a0b279d82f766091235f24f429405da6
(cherry picked from commit 86d9ec5d5932557ade18e7893cc2b8f564b5b2d8)