Warn on driver detach errors
There are cases on the detach process where an error is not considered a
failure, so the exception is ignored.

The problem is that we currently may not leave any trace in the logs
that an error actually happened.

This patch adds a simple warning log message to ensure that at least
there is some kind of notification about it.

Change-Id: Ib23debd0a2ec09c588ad122277f97a04ddb70bff