nvmeof: Support new connection properties
The os-brick nvmeof connector supports  2 different connection
properties formats, the original one and a later one that added support
to multiple portals as well as replicated devices for a software RAID.

Targets that inherit from the nvmeof base target (spdk and nvmet)
supported the original connection properties format, but it was decided
that new features, such as multipathing, would only be added to os-brick
in the new format code path.

This patch adds support to the nvmeof target class (and specifically to
the nvmet target, though it should work for other as well) for the new
connection properties format to enable support for future features.

Support for the old connection properties has been maintained, and is
still the default, since we need an easy way to exert the old code path
in os-brick to ensure that the code still works.

Configuration option ``nvmeof_conn_info_version`` is used to select what
version of the connection properties the nvmet target should use. With
version 1 being the old format and version 2 the new format. Defaults to
the old format to preserve existing behavior.

Change-Id: If3f7f66a5cd23604cc81a6973304db9f9018fdb3