db: Increase timeout for migration tests
We're seeing quite a few timeout failures on the following tests in
'cinder.tests.unit.db.test_migrations':

- TestModelsSyncMySQL.test_models_sync
- TestMigrationsWalkMySQL.test_walk_versions

Currently, this is mainly affecting the AArch64 CI but as we saw in
nova, it's possible for this to have an impact on other underpowered CI
hosts. Like nova change I82b9a064d77251945ff1ae99d7049f367ddde92e, we
now set 'TIMEOUT_SCALING_FACTOR' for these tests to 4, meaning these
tests will now get 240 seconds (or 4 minutes) to execute (we set
OS_TEST_TIMEOUT to 60 in 'tox.ini'). As discussed in that change, there
are other ways we could address this but this seems the most sensible
approach to take.

Change-Id: Iaab4984ed5d4cc8d02ef934a0c6f5980a0fb4687
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>