LVM nvmet: Add support for shared subsystems
LVM target drivers usually only support unique subsystems/targets, so a
specific subsystem/target is created for each volume.

While this is good from a deployment point of view, it is insufficient
from a testing perspective, since it limits the code paths that can be
tested in os-brick.

Being able to test these 2 different paths in os-brick is very
important, because the shared case usually present very particular
issues: Leftover devices caused by race conditions between nova and
cinder, premature subsystem/target disconnection, not disconnecting
subsystem/target, etc.

Thanks to this patch we'll be able to increase the testing possibilities
of the NVMe-oF os-brick connector to cover combinations of:

- Different connection properties formats: old & new
- Different target sharing: shared & non shared

Change-Id: I396db66f72fbf1f31f279d4431c64c9004a1a665