Stop removing .pyc files for unit test runs
We no longer need to do this.  If you run
"tox -e py38", "tox -e py39", and "tox -e py310",
you will see:

$ ls -l cinder/__pycache__/context*
 cinder/__pycache__/context.cpython-310.pyc
 cinder/__pycache__/context.cpython-38.pyc
 cinder/__pycache__/context.cpython-39.pyc

See PEP-3147 for details on this.

This is different from the older behavior, which
would generate cinder/context.pyc and collide between
different versions of Python.  In the versions of
Python that we support, no such collisions happen.

Change-Id: I1262c3ab355dfeabfa57184b6496d7fddd1e8055