add netapp copyoffload provider location
By volume creation the volume database entry is created with the field provider_location = null.
When cinder tries to register an image-cache from the new volume it gets the volume information
from the database and can not derive the mountpoint for "None".
This fix will set the provider location to prevent the error mentioned in the issue.

Closes-Bug: 1965952
Change-Id: I0fe11542d2d5440d1048d388e7a7b09055cf3af8