PowerFlex driver - fix the display of the incorrect volume size on volume or snapshot creation.
Now the user can see the size that would initially set by
him when user creates volume via powerflex driver.

This is happening because the powerflex storage creates
volumes with a size multiple of 8, but the cinder doesn't
expect the powerflex driver to return a rounded size.
As a result, the user will see not the size that he set
initially  when creating the volume, but in the storage
will be volume with the size of multiple 8.

At the end of the create volume process, if the volume
size is not equal to the original size, then extend volume
and PowerFlex driver will not return size multiply of 8.

Closes-Bug: #1968164
Change-Id: Ib0334d2e1be80d1a20c26b4435076187b0e43d40