[IBM DS8000] Fixed Detach for multi-attach volumes
During detach operation for Multi attach type volumes
the first detach from the host is successful.
In the second detach operation the volume remains in
detaching state.

Fixed the above issue by adding a check for multi-attach
volumes and get the host mapping info of a volume.
If the volume is attached to multiple hosts, then skip
the delete host operation and return.

Closes-bug: #1951046

Change-Id: I8f4586292d1310d6f7bd9fd22b98715f3414d23c