[SVf]:Fix multiple lsvdisk calls for GMCV create volume operation
[Spectrum Virtualize family] During creation of GMCV type volume
there are multiple lsvdisk calls made for checking the existence
of source change volume, aux volume and change aux volume.
Along with these calls there is a lssystem call which can be
optimized.

This patch optimizes these lsvdisk and lssystem calls which in
turn reduces the computational time for creating GMCV volumes.

Closes-Bug: #1961548

Change-Id: Id00cda88132184e66fbf6c03965f1de30cf02dcf