NetApp NFS: Clone image using copy file operation
The copy offload tool is no longer available available, so this patch adds
an alternative approach using file copy operations to the scenarios where
the tool was required. During the clone image operation, if the source
volume is not available in the same pool or cache, the copy file operation
is called to perform the action.

The copy offload tool will keeping working and it will be removed in the
next release (Antelope). There patch [1] deprecates the tool.

[1] https://review.opendev.org/c/openstack/cinder/+/847733

Implements: blueprint netapp-nfs-copy-offload-image
Change-Id: I4e2163661b913800a5a1179a5899ec1384f8c167