Update Volume Delete rejection message
When a user tries to delete a volume in "awaiting-transfer" state ,
the error message received does not include the "awaiting-transfer"
state in invalid states of Volume for deletion.
This leads to the user believing the volume-delete request is valid
and is not able to debug the reason for failure of the volume deletion
for volume present in "awaiting-transfer" state.

Closes-Bug: #1971603

Change-Id: I78915c332169b26ffb2b97310efedec65bc25e4d