Add missing extend_target driver method
The base target driver class is missing the extend_target method which
is called from the LVM driver.

Since it was missing there some drivers, such as the nvmeof base class,
also missed it when they were implemented.

This patch adds a base implementation that does nothing, since this is
usually the right thing to do for most drivers.

Change-Id: If009af0a385d9203bd74e42e822dca299a119ca7