Increase swap size to 4GB
This is to address issues in the cinder-plugin-ceph-tempest* jobs
where a significant number of builds are failing.  Symptoms are a
lot of server fault errors, identity error: unexpected status
code 500, and db connection errors in the tempest log.  On these
jobs, syslog shows multiple
   oom_reaper: reaped process xxxxxxx (mysqld)
lines.  Occasionally ceph-osd or ceph-mons are also reaped, but
it's pretty consistently mysqld.

Default swap size is currently 1G.  Increase to 4GB was determined
experimentally.  (See other patches on topic: increase-swap.)

Change-Id: I3c08f363f5b74e1fe7900f4398bd2cb6df3a80d7