Remove single-use test function
Follow-up from change I3c0e365f5dc3, where Rajat noticed that
the patch used a static test method that was used nowhere else.
Refactored the code to remove the static function, which makes
the test more readable, because now it is obvious that the test
is checking a refreshed backup object.

Change-Id: I16bdc81df4b1b01c7918a3da4024081079bcc969
Related-bug: #1965847