Fix Infinidat driver to backup attached volume
For some historical reason, the Infinidat Cinder driver does not
allow to backup an attached volume. This makes it impossible to
create a backup for most use cases. As a possible solution we can
create and use snapshots as a backup source.

This patch implements attach/detach for snapshots and the ability
to use snapshots to perform non-disruptive backups.

Added functions:
* initialize_connection_snapshot()
* terminate_connection_snapshot()
* create_export_snapshot()
* remove_export_snapshot()
* backup_use_temp_snapshot()

And changed functions:
* create_volume_from_snapshot()
* create_cloned_volume()

remove the duplicated code and use the temporary snapshot
as the source for creating the cloned volume.

Closes-bug: #1983287
Change-Id: I0e1534c6015eaa8baef48b7878324625a4cc9f1a
Signed-off-by: Alexander Deiter <adeiter@infinidat.com>