[docs] Update PTL info in contributor guide
Some recent emails have made me realize that the basic contributor
doc doesn't have info about where you can find out who the current
PTL is.  This patch moves the PTL info from the bottom of the doc
to a more visible place, and adds a brief description of what the
PTL does and who the electorate is.

Change-Id: Ib1fe4fba3d4911222b2c9393700c411d08fe2805