[Pure Storage] Fix issue with loss of replicated array
When an array on either side of a replicated configuration loses
connectivity to Cinder this can cause the driver to stop working.

This patch resolves that issue by validating all arrays in a
replicated pair exist and raising a warning if that is not
the case.

Closes Bug: #1969784
Co-Authored-By: Keerthivasan <ksuresh@purestorage.com>

Change-Id: I3256875a33a30560b834b4c8a0c6bb1b5edf69fa
(cherry picked from commit deeee2b363c4f6acf0e5cfaa6b2b04b6eb6c8362)