Doc: Fix way to update default quota value for a new project
This commit adds a way to update a default quota value for a new
project when ``use_default_quota_class = True`` which is the default
setting. With the default setting, the original way which is updating
cinder.conf doesn't update the default quota value in the database, and
the value in cinder.conf is only used in an initialize phase. So, we
need to use CLI to update default quota values in that case.

Change-Id: I01114211010817fbc40e959c3db9045958c78d06