Revert "Cleanup code duplication in cinder.cmd.backup module"
This reverts commit f6c60d1d9c79a8048df7cc63fddb2b081eac430f.

The original patch referred as duplicated code the part where the
service uses a single process when there is no additional processes
requested (backup_workers == 1), which is the default.

That code made the cinder-backup service always run as a child process
even when there was only going to be 1 process.

This adds 30MB of extra RAM consumption by the service, which amounts to
a 30% unnecessary memory increase when not doing operations after
applying changes from Ic9030d01468b3189350f83b04a8d1d346c489d3c.

Change-Id: I43a20c8687f12bc52b014611cc6977c4c3ca212c