db: Use cinder.db.api directly
`cinder/db/__init__.py` was importing `*` from `cinder.db.api`. This
meant that any time any code anywhere within the `cinder.db` package was
imported then `cinder.db.api` was too, leading to a cascade of imports
that may not have been desired. Also, in general, code in `__init__.py`
is a pain.

Therefore, this change adjusts code that so that either:

* `cinder.db.api` is used directly, or
* `cinder.db.api` is imported as `db`

In either case, the functionality remains the same.

The primary goal of this change is to make it possible to import the
model files without having to import the DB API. This looks like a very
large change, but it is essentially adjusting package names.

Note that nova did this some time back in change
Ic1fd7c87ceda05eeb96735da2a415ef37060bb1a.

Change-Id: Ie73c74c0eeeda185a8d648444a88f7ddcbdf83b4
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>