Serialize message_* properties of RequestContext
Change Idc00b125b33b added the ability to store and retrieve user
message data in the request context, but it neglected to add code to
make sure they would persist across serialization and deserialization
of the request context object, as happens in the create-backup
workflow.  As a result, when a user message needed to be created to
describe an error condition in the backup driver, instead we'd get an
exception raised: "TypeError: 'NoneType' object is not subscriptable".

This patch fixes the issue by making sure the message_* properties
persist across serialization/deserialization of RequestContext objects.

Closes-Bug: 1978729
Change-Id: Ibdaaf39abafbae6bfcb9fdf9fb7a55d968ad9f11
Signed-off-by: Jesper Schmitz Mouridsen <jesper@schmitz.computer>