Log when waiting to acquire coordinator lock
We currently have no log entry when we start trying to acquire a lock.

In general this is ok, but there are cases where it can be problematic
for example if we have a deadlock situation or if a lock takes a very
long time to be acquired.

In those scenarios looking at the logs we would see the operation
proceed normally and suddenly go completely silent without knowing that
is waiting for a lock to be freed.

This patch adds a debug log message right before trying to acquire the
lock so we can detect those situations.

The original patch was using f-strings in debug logs, which we should
avoid and newer flake8 versions is raised as error G004, so this patch
also includes Change-Id: Ia87a0114738c82b2af2391bd5b4c2acb25db71fd that
fixes this issue.

Change-Id: Ida70e9b6098b2e0426cc479811e08a164ed24899
(cherry picked from commit a423d9f9d5672fc92de2c6aae35f0c99140c408d)